% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulacao.R
\name{calcularInfiltracoes}
\alias{calcularInfiltracoes}
\title{Calcular Infiltrações}
\usage{
calcularInfiltracoes(dados)
}
\arguments{
\item{dados}{Data frame para realizar tarefa}
}
\value{
Dada frame com os dados calculados
}
\description{
Pega o dados da tabela e calcula em ordem decrescente os andares com maiores
incidências de infiltração
}
\author{
Fazendaaa
}
\keyword{internal}
