% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulacao.R
\name{calcularMedidas}
\alias{calcularMedidas}
\title{Calcular Tamanho}
\usage{
calcularMedidas(dados)
}
\arguments{
\item{dados}{Data frame para realizar tarefa}
}
\value{
Dada frame com os dados calculados
}
\description{
Pega o dados da tabela e calcula medidas de posição e disperção
}
\author{
Fazendaaa
}
