% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulacao.R
\name{calcularPosicao}
\alias{calcularPosicao}
\title{Calcular Posição}
\usage{
calcularPosicao(dados)
}
\arguments{
\item{dados}{Data frame para realizar tarefa}
}
\value{
Dada frame com os dados calculados
}
\description{
Pega o dados da tabela e calcula as medidas de posição como:
\itemize{
\item
\item
\item
}
}
\author{
Fazendaaa
}
\keyword{internal}
