% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulacao.R
\name{calcularRachaduras}
\alias{calcularRachaduras}
\title{Calcular Rachaduras}
\usage{
calcularRachaduras(dados)
}
\arguments{
\item{dados}{Data frame para realizar tarefa}
}
\value{
Dada frame com os dados calculados
}
\description{
Pega o dados da tabela e calcula em ordem decrescente os andares com maiores
incidências de rachaduras
}
\author{
Fazendaaa
}
\keyword{internal}
