% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulacao.R
\name{calcularTamanho}
\alias{calcularTamanho}
\title{Calcular Tamanho}
\usage{
calcularTamanho(dados)
}
\arguments{
\item{dados}{Data frame para realizar tarefa}
}
\value{
Dada frame com os dados calculados
}
\description{
Pega o dados da tabela e cacula com o tamanho de cada bloco em uma nova
coluna
}
\author{
Fazendaaa
}
\keyword{internal}
