% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulacao.R
\name{calcularTotal}
\alias{calcularTotal}
\title{Calcular Total}
\usage{
calcularTotal(dados)
}
\arguments{
\item{dados}{Data frame para realizar tarefa}
}
\value{
Dada frame com os dados calculados
}
\description{
Pega o dados da tabela e cacula o tamanho total de blocos e com o tamanho de
cada um
}
\author{
Fazendaaa
}
