% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visulizacao.R
\name{gerarVisualizacao}
\alias{gerarVisualizacao}
\title{Gerar Visualização}
\usage{
gerarVisualizacao(dados)
}
\arguments{
\item{dados}{Data frame para realizar tarefa}
}
\value{
Box plot
}
\description{
Gera a visualização em box plot com o tamanho de cada bloco
}
\author{
Fazendaaa
}
