% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arquivo.R
\name{lerArquivo}
\alias{lerArquivo}
\title{Ler Arquivo}
\usage{
lerArquivo()
}
\value{
O arquivo em formato Data Frame
}
\description{
Lê o arquivo e o retorna
}
\section{Observações}{

Toda e qualquer modificação do arquivo deverá ser feita aqui antes de ser
processado
}

\author{
Fazendaaa
}
