% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arquivo.R
\name{salvarArquivo}
\alias{salvarArquivo}
\title{Salvar Arquivo}
\usage{
salvarArquivo(dados)
}
\arguments{
\item{dados}{Data frame estrutura que será salva}
}
\description{
Salva os dados em um arquivo xlsx
}
\section{Observações}{

Toda e qualquer modificação dos dados no tocante a estrutura xlsx deverá ser
feita aqui antes de ser salvo
}

\author{
Fazendaaa
}
