% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulacao.R
\name{verificarOcorrencias}
\alias{verificarOcorrencias}
\title{Calcular Ocorrências}
\usage{
verificarOcorrencias(dados)
}
\arguments{
\item{dados}{Data frame para realizar tarefa}
}
\value{
Dada frame com os dados calculados
}
\description{
Pega o dados da tabela e calcula se infiltrações e rachaduras aparcem mais:
\itemize{
\item Nos primeiros andares
\item Nos últimos
\item Não há relação
}
}
\author{
Fazendaaa
}
